+++ Cargo.toml
@@ -1,5 +1,15 @@
-[workspace]
-members = [
-    "crates/app",
-    "crates/downloader",
-]
+[package]
+name = "bilibili-downloader"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+tokio = { version = "1", features = ["full"] }
+futures = { version = "^0.3.23" }
+clap = "2.34.0"
+reqwest = { version = "^0.11", features = ["json", "deflate"] }
+serde_json = "1.0"
+anyhow = "1.0"
+thiserror = { version = "^1.0.32" }
