--- crates/app/Cargo.toml
@@ -1,12 +1,0 @@
-[package]
-name = "app"
-version = "0.1.0"
-edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-downloader = { path = "../downloader" }
-tokio = { version = "1", features = ["full"] }
-futures = { version = "^0.3.23" }
-clap = "2.34.0"
