--- crates/downloader/Cargo.toml
@@ -1,12 +1,0 @@
-[package]
-name = "downloader"
-version = "0.1.0"
-edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-reqwest = { version = "^0.11", features = ["json", "deflate"] }
-serde_json = "1.0"
-anyhow = "1.0"
-thiserror = { version = "^1.0.32" }
