+++ src/download/mod.rs
@@ -1,0 +1,5 @@
+mod downloader;
+pub use downloader::Downloader;
+
+mod model;
+pub use model::Video;
