+++ src/download/model.rs
@@ -1,0 +1,9 @@
+#[derive(Debug, Clone)]
+pub struct Video {
+    pub bv: String,
+    pub cid: String,
+    pub url: String,
+    pub title: String,
+    pub format: String,
+    pub content_lenth: u64,
+}
